test(upskill): split monolithic test into per-module files mirroring upskill/ modules
